# docs/swagger.yml
openapi: '3.0.2'
info:
  title: RecyThing
  version: '1.0'
  description: "API Documentation Capstone Project RecyThing Kelompok 2"
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /register:
    post:
      tags:
        - authentication
      summary: Register a new user
      description: Register new account
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        '201':
          description: Register Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: user successfully register! OTP sent to your email
                  data:
                    type: object
                    properties:
                      user_id:
                        type: string
                        example: U00001
                      name:
                        type: string
                        example: U00001
                      email:
                        type: string
                        example: john.doe@gmail.com
                      is_verified:
                        type: boolean
                        example: false
        '400':
          description: Register Failed (validation error / email already exists)
  /verify-otp:
    post:
      tags:
        - authentication
      summary: Verify OTP
      description: Verify the OTP sent to the user email after registration.
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: The email of the user
                  example: john.doe@gmail.com
                otp:
                  type: string
                  description: The OTP sent to the user's email
                  example: 678901
      responses:
        '200':
          description: OTP successfully verified, user registration complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: OTP successfully verified. Registration complete!
        '400':
          description: Invalid OTP or user ID
  /resend-otp:
    post:
      tags:
        - authentication
      summary: Resend OTP
      description: Resend OTP sent to the user email.
      operationId: resendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: The email of the user
                  example: john.doe@gmail.com
      responses:
        '200':
          description: OTP successfully verified, user registration complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: new otp sent to your email!
        '400':
          description: Invalid email or user already verified
  /login:
    post:
      tags:
        - authentication
      summary: Login User
      description: Login account
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Login Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: login successfully!
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        example: john.doe@gmail.com
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiVVNSMDAwMSIsInJvbGUiOiJ1c2VyIiwiZXhwIjoxNzE2ODI3NDE1fQ.o4rXzRLdWGS3iqtznZg9hhJyOklWyrNUA0MgOmUkfis
        '400':
          description: Login Failed (invalid email / password)
  /admin/login:
    post:
      tags:
        - authentication
      summary: Login Super admin / Admin
      description: Login account
      operationId: loginAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          description: Login Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: login successfully!
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        example: john.doe@gmail.com
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiVVNSMDAwMSIsInJvbGUiOiJ1c2VyIiwiZXhwIjoxNzE2ODI3NDE1fQ.o4rXzRLdWGS3iqtznZg9hhJyOklWyrNUA0MgOmUkfis
        '400':
          description: Login Failed (invalid email / password)
...
  /article/{articleId}:
    get:
      tags:
        - articles
        - manage articles
      summary: Get article data by id
      description: Endpoint user to get article by id
      operationId: getArticleById
      parameters:
        - name: articleId
          in: query
          required: true
          schema:
            type: string
            # default: ART0001
          description: Article ID
          example: ART0001
      security:
        - Bearer: []
      responses:
        '200':
          description: Get data article
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message: 
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      schema:
                        $ref: '#/components/schemas/Article'
        '400':
          description: Invalid param request
        '404':
          description: Article not found
    put:
      tags:
        - manage articles
      summary: Edit article
      description: Endpoint admin to edit an existing article.
      operationId: editArticle
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
          description: ID of the article to be edited
          example: ART0001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
      security:
        - Bearer: []
      responses:
        '200':
          description: Article successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: article updated!
        '404':
          description: Article not found
        '400':
          description: Invalid request data 
    delete:
      tags:
        - manage articles
      summary: Delete article
      description: Endpoint admin to delete an article.
      operationId: deleteArticle
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
          description: ID of the article to be deleted
          example: ART0001
      security:
        - Bearer: []
      responses:
        '200':
          description: Article successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: article successfully deleted!
        '404':
          description: Article not found
...
  /videos/categories:
    get:
      tags:
        - manage videos
      summary: Get All Categories
      description: Get All Categories
      operationId: getAllCategories
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        content_categories:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "tips"
                        waste_categories:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "berbahaya"
                      
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: error get all categories
...
components:
  schemas:
...
    UserRegister:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@gmail.com
        password:
          type: string
          example: password@123
    UserLogin:
      type: object
      properties:
        email:
          type: string
          example: john.doe@gmail.com
        password:
          type: string
          example: password@123
...
    Article:
      type: object
      properties:
        id:
          type: string
          example: ART0001
        author:
          type: object
          properties:
            id:
              type: string
              example: AD0001
            name: 
              type: string
              example: John Doe
            image_url:
              type: string
              example: http://example.com/image.png
        title:
          type: string
          example: Bahaya Limbah Elektronik dan Cara Pembuangan yang Bertanggung Jawab
        description:
          type: string
          example: Pelajari tentang bahaya lingkungan dan kesehatan dari limbah elektronik dan temukan metode pembuangan yang aman.
        thumbnail_url:
          type: string
          format: uri
          example: http://example.com/image.png
        waste_categories:
          type: array
          items:
            type: string
          example: ["elektronik", "berbahaya"]
        content_categories:
          type: array
          items:
            type: string
          example: ["edukasi", "kampanye"]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ArticleSection'
    ArticleSection:
      type: object
      properties:
        title:
          type: string
          example: Apa itu Limbah Elektronik?
        description:
          type: string
          example: Limbah elektronik mencakup barang elektronik bekas seperti ponsel, komputer, peralatan rumah tangga, dll.
        image_url:
          type: string
          format: uri
          example: http://example.com/image.png
...